Added .idea settings and detailed README
The README mentions the entire study plan I'll follow for the next 6 weeks. I have kept it a bit vague and not very tight in terms of time to allow for experimentation and breaks in between.

Going home, so might not be able to commit as much time as I want :(