Added docker configuration for deploying CDK stack
This is completely overkill and only for the sake of the demo.
You don't need to build the deployment in a docker build and then use a separate docker image for deploy.
This is not what docker should be used for, only what it can be used for.
This was the only project I had at hand, else ideally, I would have used a simple express application, which would have let me incorporate a health check as well